added decoder factory
